chore(lint): clear the golangci-lint baseline and turn the gate on
147 findings to zero, and the gate is now in quality:verify and in CI.

The count was the first thing to fix. golangci-lint caps output at three
issues per linter and three of the same kind, so every number taken from
it was the cap: fixing three bodyclose findings revealed the next three,
and the total sat around 80 while real progress happened underneath. A
baseline measured against that would have gone green on a fraction of the
work.

Two real bugs came out of it.

`bb search prs --repo X` swallowed its error. `projectKey, slug, err :=
reposel.Resolve(...)` shadowed the outer err, so service.List assigned its
failure to a variable that fell out of scope and the check after the
branch read a nil. A failed search returned an empty list and exit 0.

The MCP audit directory was created 0755. It records which repositories an
agent touched and what it did to them, and ADR-062 makes it a compliance
artefact; one every local account can read is a weaker one. Now 0750.

Three int-to-int32 conversions rejected rather than wrapped. `bb auth
token create --expiry-days 2147483648` reached the server as a negative
expiry; a bulk policy could set an inactivityWeeks that configured the
opposite of what it said; a pull request version outside int32 rebased
against a different version than the caller named. These fail invocations
that previously "succeeded", which is why they belong in a major.

Command output that never lands is no longer reported as success.
`bb pr list > file` on a full disk printed nothing and exited 0: every
fmt.Fprint discarded its error, correctly, because a command cannot report
that stdout is broken by writing to stdout. internal/cli/outwriter wraps
stdout once at the root and cmd/bb checks it after the command returns,
where an exit code is still available. A closed pipe stays a success --
`bb pr list | head` is ordinary usage, not a failure.

Two dead duplicates left by the ADR-032 modularization, both with no
callers: normalizeAccessKeyIDs, an int32 bounds validator duplicated
inline in branch.go, and resolveCommentTarget. The live copy of the second
had no test of its own, so the test moved rather than being deleted; it
now guards code that runs. Four more dead functions came from the pull
request task commands removed in #386, and one was named in three test
comments claiming to exercise it.

35 tests captured command output and asserted nothing about it. 33 were
--dry-run invocations, where the preview is the entire product -- a test
that discards it cannot tell a real preview from an empty one, which is
the shape of #481. They now assert a preview was produced. The --blocker
filter on `comment list` reads a different endpoint, so asserting its
output is what separates the flag being honoured from the flag being
accepted and ignored, which is #476's complaint across 113 commands. None
of the 35 warranted masking.

Where the linter was wrong, the reasoning is recorded at the site rather
than swept into config: three identity comparisons errors.Is would have
weakened, five subprocess launches that are a fixed binary with arguments
passed as a slice, and a tar mode masked to permission bits.

Two exclusions are scoped and explained. staticcheck's QF quickfix
category is off -- 32 editor refactors with no behaviour change; writing
that as ["all", "-QF*"] silently enables the ST checks the default
excludes, turning 44 findings into 200, so the config carries the default
list with -QF* appended. gosec's file permission checks are off for tools/
only, where they wanted 0600 on generated artifacts that are committed and
meant to be world-readable; they stay on for cmd/ and internal/, which is
where they found the audit directory.

govulncheck arrived with the linter and immediately found five standard
library vulnerabilities this code calls, because the pinned toolchain was
five patches behind. go.mod now carries toolchain go1.26.6.

Closes #457

+++ internal/cli/cmd/repo/comments.go
@@ -57,7 +57,7 @@ func commentIDString(comment openapigenerated.RestComment) string {
 	if comment.Id == nil {
 		return "?"
 	}
-	return strconv.FormatInt(int64(*comment.Id), 10)
+	return strconv.FormatInt(*comment.Id, 10)
 }
 
 func formatCommentSummary(comment openapigenerated.RestComment) string {
